ARPuzzleControl script started
The ARPuzzleControl script was started and canvas content was added to the scene as well. The canvas is not active until after the player has clicked or tapped the coin, after which the coin will disappear an the canvas will appear. The button currently does nothing. Style and other artistic content will have to be added to the canvas in the scene.